# backend/openapi_v3.yaml
openapi: 3.0.1
info:
  title: 团队心流引擎 (TeamFlow Engine) API
  description: 为《团队心流引擎》应用提供后端服务的RESTful API。该API支持团队管理、每日状态签到、数据可视化和AI分析等核心功能。
  version: 1.2.1
...
paths:
...
  /auth/register/:
    post:
      summary: 用户注册
      deprecated: false
      description: ''
      operationId: authRegister
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
        required: true
      responses:
        '201':
          description: 用户创建成功
          headers: {}
        '400':
          description: 无效的输入数据
          headers: {}
      security: []
  /auth/login/:
    post:
      summary: 用户登录
      deprecated: false
      description: ''
      operationId: authLogin
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
        required: true
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
          headers: {}
      security: []
  /auth/token/refresh/:
    post:
      summary: 刷新Access Token
      deprecated: false
      description: ''
      operationId: refreshToken
      tags:
        - Authentication
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh
              properties:
                refresh:
                  type: string
        required: true
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
          headers: {}
      security: []
...
components:
  schemas:
...
    TokenPair:
      type: object
      properties:
        refresh:
          type: string
        access:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers: []
security: []